+++ examples/two-pages/src/App/App.hs
@@ -268,6 +268,7 @@ twoPageRequestPolicy =
       trustForwardedHeaders = False,
       requestHeadLimits = unboundedRequestHeadLimits,
       requestTransportLimits = warpDefaultRequestTransportLimits,
+      requestConcurrencyLimit = Nothing,
       corsPolicy = defaultCorsPolicyConfig,
       responseSecurityHeaders = defaultResponseSecurityHeadersConfig
     }