+++ packages/harch-web/src/HarchWeb/Site.hs
@@ -250,6 +250,7 @@ defaultSiteRequestPolicy =
       trustForwardedHeaders = False,
       requestHeadLimits = unboundedRequestHeadLimits,
       requestTransportLimits = warpDefaultRequestTransportLimits,
+      requestConcurrencyLimit = Nothing,
       corsPolicy = defaultCorsPolicyConfig,
       responseSecurityHeaders = defaultResponseSecurityHeadersConfig
     }